# api-docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BannerListResponse": {
        "properties": {
          "banners": {
            "items": {
              "$ref": "#/components/schemas/app__api__routers__admin_landings__BannerResponse"
            },
            "title": "Banners",
            "type": "array"
          }
        },
        "required": [
          "banners"
        ],
        "title": "BannerListResponse",
        "type": "object"
      },
      "BannerOrderRequest": {
        "properties": {
          "banner_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Banner Ids",
            "type": "array"
          }
        },
        "required": [
          "banner_ids"
        ],
        "title": "BannerOrderRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LandingBlockOrderRequest": {
        "description": "Full current block-id sequence, reordered only by the dashboard.",
        "properties": {
          "block_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Block Ids",
            "type": "array"
          }
        },
        "required": [
          "block_ids"
        ],
        "title": "LandingBlockOrderRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "The same session token supplied as `Authorization: Bearer <token>`, for clients that cannot hold cookies. Accepted everywhere the cookie is.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Session cookie issued by `POST /api/auth/login`. `HttpOnly`, `SameSite=Strict`, and `Secure` outside local development, so a browser sends it automatically and page scripts cannot read it. This is the transport browser clients should use.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "summary": "Cash-on-delivery commerce API for a single merchant: public landing delivery and checkout, plus the authenticated Administrator surface.",
    "title": "Easy Funnel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/landings/{landing_id}/banners/order": {
      "put": {
        "description": "Apply an explicit full ordering of the landing's banners.",
        "operationId": "reorder_banners_api_admin_landings__landing_id__banners_order_put",
        "parameters": [
          {
            "in": "path",
            "name": "landing_id",
            "required": true,
            "schema": {
              "title": "Landing Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BannerOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BannerListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Reorder Banners",
        "tags": [
          "landings"
        ]
      }
    },
...
    "/api/admin/landings/{landing_id}/blocks/order": {
      "patch": {
        "description": "Reorder components within their existing slots without editing content.",
        "operationId": "reorder_landing_blocks_api_admin_landings__landing_id__blocks_order_patch",
        "parameters": [
          {
            "in": "path",
            "name": "landing_id",
            "required": true,
            "schema": {
              "title": "Landing Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LandingBlockOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandingBlockListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Reorder Landing Blocks",
        "tags": [
          "landings"
        ]
      }
    },
...
    "/api/admin/orders": {
      "get": {
        "description": "List orders with filters (ANDed together).",
        "operationId": "list_orders_api_admin_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "in": "query",
            "name": "landing_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Landing Id"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Orders Api Admin Orders Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Orders",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/admin/orders/export.csv": {
      "get": {
        "description": "Export orders to CSV with formula-injection neutralization.",
        "operationId": "export_orders_csv_api_admin_orders_export_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "in": "query",
            "name": "landing_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Landing Id"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Export Orders Csv",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/admin/orders/{order_id}": {
      "get": {
        "description": "Get order detail with fraud flags.",
        "operationId": "get_order_api_admin_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Order Api Admin Orders  Order Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/admin/orders/{order_id}/fulfillment": {
      "patch": {
        "description": "Edit delivery data before a successful external synchronization.",
        "operationId": "update_order_fulfillment_api_admin_orders__order_id__fulfillment_patch",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderFulfillmentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Order Fulfillment Api Admin Orders  Order Id  Fulfillment Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update Order Fulfillment",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/admin/orders/{order_id}/transition": {
      "post": {
        "description": "Transition order status (legal graph only).",
        "operationId": "transition_order_api_admin_orders__order_id__transition_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Transition Order",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
...
}